+++ pkg/annotation/annotation.go
@@ -3,5 +3,6 @@ package annotation
 const (
 	Docs                    = "giantswarm.io/docs"
 	InstanceID              = "aws-operator.giantswarm.io/instance"
+	LegacyAwsCniPodCidr     = "aws-operator.giantswarm.io/legacy-aws-cni-pod-cidr"
 	MachineDeploymentSubnet = "machine-deployment.giantswarm.io/subnet"
 )